create listing request file
